Sort permission call to work around /permissions API bug
For large workspaces (4000+ users in our case), we have issues with the paged queries to the /permissions API: Some users are returned as duplicates, while others are dropped entirey. Adding a sort parameter works around this issue (but I would of course appreciate if the problem was rather fixed internally). Thanks!